pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-suite"
version = "0.1.0"
description = "Monorepo of four CLI tools: timeline-maker, quote-maker, proposal-maker, project-maker."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "project-suite contributors" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0.1",
    "pydantic>=2.6",
    "openpyxl>=3.1.2",
    "python-docx>=1.1",
    "markdown-it-py>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
]

[project.scripts]
timeline-maker = "timeline_maker.cli:app"
quote-maker = "quote_maker.cli:app"
proposal-maker = "proposal_maker.cli:app"
project-maker = "project_maker.cli:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"src/timeline_maker/cli.py" = ["B008"]
"src/quote_maker/cli.py" = ["B008"]
"src/proposal_maker/cli.py" = ["B008"]
"src/project_maker/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
